+++ ethereum/forge-test/relayer/WormholeRelayerGovernance.t.sol
@@ -173,7 +173,6 @@ contract WormholeRelayerGovernanceTests is Test {
                 == toWormholeFormat(address(wormholeRelayer2))
         );
     }
-    
     function testUpgradeContractToItself() public {
         address payable myCoreRelayer = payable(
             address(helpers.setUpCoreRelayer(wormhole, address(relayProvider)))
